requirements-onnx.txt
torch
onnx
onnxruntime
numpy
pillow
logging
